Task Id: Update form control ability and cleanup

+++ TaskHub/src/main/java/taskhub/action/uibean/sa/SalesBean.java
@@ -1,35 +1,40 @@
 package taskhub.action.uibean.sa;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.enterprise.context.ConversationScoped;
+import javax.faces.model.SelectItem;
 import javax.inject.Inject;
 import javax.inject.Named;
 
 import taskhub.action.service.sa.SalesService;
 import taskhub.action.uibean.AbstractBean;
+import taskhub.persistence.constant.Inventory;
+import taskhub.persistence.constant.ItemType;
+import taskhub.persistence.constant.Service;
+import taskhub.persistence.constant.UomType;
 import taskhub.persistence.entity.sa.Sa_order_new_det;
 import taskhub.persistence.entity.sa.Sa_order_new_hdr;
-import taskhub.persistence.entity.sa.Sa_order_ost_det;
 import taskhub.persistence.entity.sa.Sa_order_ost_hdr;
 
 @SuppressWarnings("serial")
 @Named
 @ConversationScoped
 public class SalesBean extends AbstractBean {
+
 	private static final String MODULE_CODE = "SA";
 
 	@Inject
 	private SalesService service;
 
 	private Sa_order_new_hdr current_Sa_order_new_hdr;
+	private Sa_order_ost_hdr current_Sa_order_ost_hdr;
 	private Sa_order_new_det selected_Sa_order_new_det;
 	
-	private Sa_order_ost_hdr current_Sa_order_ost_hdr;
-	private Sa_order_ost_det selected_Sa_order_ost_det;
-
-	private List<Sa_order_new_hdr> sa_order_new_hdrs;
 	
+	private List<Sa_order_new_hdr> sa_order_new_hdrs;
+
 	private List<Sa_order_ost_hdr> sa_order_ost_hdrs;
 
 	public void create() {
@@ -57,18 +62,23 @@ public class SalesBean extends AbstractBean {
 	}
 
 	public void deleteSa_order_new_det() {
-		this.service.deleteSa_order_new_det(this.selected_Sa_order_new_det);
+		if (this.selected_Sa_order_new_det != null) {
+			this.service.deleteSa_order_new_det(this.selected_Sa_order_new_det);
+		}
 	}
 
 	public void revise() {
-
+		if (this.current_Sa_order_ost_hdr != null) {
+			final Sa_order_new_hdr sa_order_new_hdr = this.service.revise(this.current_Sa_order_ost_hdr);
+			this.current_Sa_order_new_hdr = sa_order_new_hdr;
+			this.current_Sa_order_ost_hdr = null;
+		}
 	}
 	
-	protected void switchToSummary() {
+	public void switchToSummary() {
 		this.current_Sa_order_new_hdr = null;
 		this.current_Sa_order_ost_hdr = null;
-		this.selected_Sa_order_new_det= null;
-		this.selected_Sa_order_ost_det = null;
+		this.selected_Sa_order_new_det = null;
 		this.sa_order_new_hdrs = null;
 		this.sa_order_ost_hdrs = null;
 	}
@@ -80,7 +90,7 @@ public class SalesBean extends AbstractBean {
 	public void setCurrent_Sa_order_new_hdr(Sa_order_new_hdr current_Sa_order_new_hdr) {
 		this.current_Sa_order_new_hdr = current_Sa_order_new_hdr;
 	}
-
+	
 	public Sa_order_new_det getSelected_Sa_order_new_det() {
 		return this.selected_Sa_order_new_det;
 	}
@@ -108,14 +118,6 @@ public class SalesBean extends AbstractBean {
 		this.current_Sa_order_ost_hdr = current_Sa_order_ost_hdr;
 	}
 
-	public Sa_order_ost_det getSelected_Sa_order_ost_det() {
-		return this.selected_Sa_order_ost_det;
-	}
-
-	public void setSelected_Sa_order_ost_det(Sa_order_ost_det selected_Sa_order_ost_det) {
-		this.selected_Sa_order_ost_det = selected_Sa_order_ost_det;
-	}
-
 	public List<Sa_order_ost_hdr> getSa_order_ost_hdrs() {
 		if (this.sa_order_ost_hdrs == null) {
 			this.sa_order_ost_hdrs = this.service.findAll_Sa_order_ost_hdr();
@@ -126,8 +128,41 @@ public class SalesBean extends AbstractBean {
 	public void setSa_order_ost_hdrs(List<Sa_order_ost_hdr> sa_order_ost_hdrs) {
 		this.sa_order_ost_hdrs = sa_order_ost_hdrs;
 	}
+
+	public List<SelectItem> getItemTypeSelectItem() {
+		final List<SelectItem> selectItems = new ArrayList<SelectItem>();
+		selectItems.add(new SelectItem(null, "Please Select"));
+		for (final ItemType it : ItemType.values()) {
+			selectItems.add(new SelectItem(it, it.name()));
+		}
+		return selectItems;
+	}
+
+	public List<SelectItem> getUomTypeSelectItem() {
+		final List<SelectItem> selectItems = new ArrayList<SelectItem>();
+		selectItems.add(new SelectItem(null, "Please Select"));
+		for (final UomType ut : UomType.values()) {
+			selectItems.add(new SelectItem(ut, ut.getLabel()));
+		}
+		return selectItems;
+	}
 	
+	public List<SelectItem> getInventorySelectItem() {
+		final List<SelectItem> selectItems = new ArrayList<SelectItem>();
+		selectItems.add(new SelectItem(null, "Please Select"));
+		for (final Inventory inv : Inventory.values()) {
+			selectItems.add(new SelectItem(inv, inv.getDesc()));
+		}
+		return selectItems;
+	}
 	
-	
+	public List<SelectItem> getServiceSelectItem() {
+		final List<SelectItem> selectItems = new ArrayList<SelectItem>();
+		selectItems.add(new SelectItem(null, "Please Select"));
+		for (final Service s : Service.values()) {
+			selectItems.add(new SelectItem(s, s.getDesc()));
+		}
+		return selectItems;
+	}
 
 }
